cli: split rewrite module into options and engine submodules
Break the monolithic rewrite/mod.rs into a pure facade (rewrite.rs)
plus options.rs (RewriteOptions, resolution, topic selection) and
engine.rs (read/select/place pipeline, summary helpers). Rename the
resolved options struct FilterOptions -> ResolvedOptions to shed the
stale filter vocabulary, and move each test group next to the code it
exercises. Uses the repo's Rust 2018 foo.rs + foo/ module convention
instead of mod.rs, and keeps inline #[cfg(test)] test modules.
